+++ packages/webcomponentsjs/ts_src/platform/es6-misc.ts
@@ -22,14 +22,15 @@ if (!Object.assign) {
     const n$ = Object.keys(source);
     for (let i = 0; i < n$.length; i++) {
       const p = n$[i];
-      // tslint:disable-next-line:no-any
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
       (target as any)[p] = (source as any)[p];
     }
   };
 
   Object.assign = function (target: object) {
+    // eslint-disable-next-line prefer-rest-params
     const args: object[] = [].slice.call(arguments, 1);
-    for (var i = 0, s; i < args.length; i++) {
+    for (let i = 0, s; i < args.length; i++) {
       s = args[i];
       if (s) {
         assign(target, s);