+++ packages/custom-elements/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+<!-- ## [Unreleased] -->
+
+## [1.4.3] - 2020-10-21
+
+- Maintenance release (no user-facing changes)
 
 ## [1.4.2] - 2020-07-20